2nd Attempt
I fixed the file IO thing I think but it won't draw anything because it fills my xs and ys arrays incorrectly the very first time throwing it off the order the rest of the time even though it is behaving correctly. I should have fixed the windows and stuff as well.